Move getRealName to utilities file
